This fully addresses issue #5 and contributes to my issue #6
The /media/drawable directory does not exist when the extension is initially
installed, yet the createOrShow() function calls fs.rmdirSync recursively,
which throws an error on the non-existing drawable directory (media exists).

I've wrapped this call up with an fs.existsSync, which will perform the
existing action if the directory exists, otherwise it calls fs.mkdirSync
with the same options as the call to fs.rmdirSync (i.e, recursive: true)

The extension will now execute successfully on an initial install without the
user having to create the subdirectory tree manually.

+++ src/extension.ts
@@ -68,12 +68,17 @@ class AppPanel {
 			// fall back to hard-coded path
 			resourcesPath = "src/main/res/drawable";
 		}
+		
 		const fse = require('fs-extra');
 		//@ts-ignore.
 		var sourceDir = path.join(rootUri?.path, resourcesPath)
 		var destinationDir = path.join(extensionUri.path, "/media/drawable")
-		fs.rmdirSync(destinationDir, { recursive: true });
-
+		if (fs.existsSync(destinationDir)) {
+			fs.rmdirSync(destinationDir, { recursive: true });
+		} else {
+			// Create cache location
+			fs.mkdirSync(destinationDir, { recursive: true });
+		}
 		// To copy a folder or file  
 		fse.copySync(sourceDir, destinationDir)
